Move helpers back to xpk.py, add helpers for resource flavors.
